+++ CHANGELOG.md
@@ -16,11 +16,11 @@ Jump to version:
 Fixes
 - Fixed control characters being included in TextInput
 - Fixed TextInputs behaving incorrectly when switching between multiline and single-line modes
+- Fixed TextInput drawing characters with the wrong width if a masking character is used
 
 ### MLEM.Ui
 Fixes
 - Fixed images not updating their hidden state properly when the displayed texture changes
-- Fixed control characters being included in TextField
 - Fixed a multiline text field's cursor not returning to the default position when the last character is removed
 
 ## 6.1.0